+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Feature) Helm Client Extension
 - (Feature) (Integration) SchedulerV2 Definition
 - (Maintenance) Proto Lint
+- (Feature) (Integration) SchedulerV2
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD